lecture-8-hw-2
• Removed unnecessary wrapper tags
• Divide huge components to sub-components (cart, header, review)
• Created "reducers" folder with formReducer file to separate reducer code from a component.
