+++ src/components/header/component.jsx
@@ -2,15 +2,13 @@ import styles from "./style.module.css";
 import classnames from "classnames";
 
 import { useTheme } from "../themeContext";
-import { useUser } from "../userContext";
 
 import { NavLink } from "react-router-dom";
 
-import UserButton from "../userButton/component";
+import LoginWrapper from "../loginWrapper/component";
 
 function Header() {
   const themeValue = useTheme();
-  const userValue = useUser();
 
   return (
     <div
@@ -23,10 +21,7 @@ function Header() {
         <NavLink to="/">Доставка еды</NavLink>
       </h1>
 
-      <div className={styles.loginWrapper}>
-        <span>{userValue.value !== null ? userValue.value : ""} </span>
-        <UserButton />
-      </div>
+      <LoginWrapper />
     </div>
   );
 }
