+++ src/components/loginWrapper/style.module.css
@@ -1,0 +1,9 @@
+.loginWrapper {
+  width: 100%;
+  height: 100%;
+
+  display: grid;
+  place-items: center;
+
+  grid-template-columns: 1fr 1fr;
+}
