+++ src/components/reviewForm/component.jsx
@@ -1,48 +1,7 @@
 import { useReducer } from "react";
-import "./style.css";
+import ReviewStars from "../reviewStars/component";
 
-const INITIAL_STATE = {
-  name: "",
-  text: "",
-  rating: 0,
-  hoveredStar: 0,
-};
-
-const formReducer = (state, action) => {
-  switch (action.type) {
-    case "CHANGE_NAME":
-      return {
-        ...state,
-        name: action.payload.name,
-      };
-
-    case "CHANGE_TEXT":
-      return {
-        ...state,
-        text: action.payload.text,
-      };
-
-    case "CHANGE_RATING":
-      return {
-        ...state,
-        rating: action.payload.rating,
-      };
-
-    case "CHANGE_HOVERED_STAR":
-      return {
-        ...state,
-        hoveredStar: action.payload.hoveredStar,
-      };
-
-    case "SAVE_REVIEW":
-      return {
-        ...INITIAL_STATE,
-      };
-
-    default:
-      return state;
-  }
-};
+import { INITIAL_STATE, formReducer } from "../../reducers/form/formReducer";
 
 function ReviewForm() {
   const [state, dispatch] = useReducer(formReducer, INITIAL_STATE);
@@ -106,37 +65,14 @@ function ReviewForm() {
         value={state.text}
         placeholder="Enter your review..."
       />
-      <div id="inputRadioContainer">
-        {[...Array(5)].map((element, index) => {
-          const currentRating = index + 1;
-
-          return (
-            <label key={index}>
-              <input
-                type="radio"
-                name="rating"
-                value={currentRating}
-                onChange={handleRatingInput}
-              />
-              <span
-                className="star"
-                style={{
-                  color:
-                    currentRating <= (state.hoveredStar || state.rating)
-                      ? "gold"
-                      : "grey",
-                }}
-                onMouseEnter={() => handleStarHover(currentRating)}
-                onMouseLeave={() => handleStarHover(0)}
-              >
-                ★
-              </span>
-            </label>
-          );
-        })}
-        <br />
-        <button onClick={handleSaveButtonClick}>Save review</button>
-      </div>
+      <ReviewStars
+        rating={state.rating}
+        hoveredStar={state.hoveredStar}
+        handleRatingInput={handleRatingInput}
+        handleStarHover={handleStarHover}
+      />
+      <br />
+      <button onClick={handleSaveButtonClick}>Save review</button>
     </div>
   );
 }
