+++ src/components/reviewStars/style.module.css
@@ -2,8 +2,17 @@
   font-size: 1.5em;
   margin: 5px;
   user-select: none;
+  cursor: pointer;
 }
 
-input[type="radio"] {
+.inputRadio {
   display: none;
+}
+
+.gold {
+  color: gold;
+}
+
+.grey {
+  color: grey;
 }
